backtesting: move backtesting to serverside
this change comes with a big performance improvement for the UI,
because the heavy work is done by the hf-data-server now, which
is a lot faster, even without clustering.

this PR also makes a few cleanups, removing "in-component-set-state"
calls and replacing them with redux based message passing.

it also removes a homegrown message passing solution using
`postMessage` which was bypassing the flux stores.

this is the first incremental step to a better backtesting and
execution model for the HF. in next steps, we will change to a
hypercore based architecture, cleaning up a lot of code.

related PRs:

https://github.com/bitfinexcom/bfx-hf-server/pull/34
https://github.com/bitfinexcom/bfx-hf-data-server/pull/22